backend/requirements.txt
fastapi
langchain
openai
pinecone-client
uvicorn
sentence-transformers